refactor(developers): extract shared CodeBlock and CopyButton components
- Create apps/developers/src/components/code/ with shared components
- Remove 164 lines of duplicated code across 3 page files
- Update AGENTS.md to clarify Sisyphus as primary agent
- Update OPENCODE-PARTNERSHIP.md with new partnership model

This refactoring improves maintainability by centralizing the code
display components used throughout the developers portal.
