+++ apps/developers/src/app/page.tsx
@@ -1,75 +1,8 @@
 "use client";
 
-import { useState } from "react";
-import { Copy, Check, ExternalLink, ChevronRight } from "lucide-react";
-import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
-import { oneDark } from "react-syntax-highlighter/dist/cjs/styles/prism";
+import { ExternalLink, ChevronRight } from "lucide-react";
 import { EditOnGitHub } from "../components/EditOnGitHub";
-import { trackCodeCopy } from "../lib/analytics";
-
-function CopyButton({ text }: { text: string }) {
-  const [copied, setCopied] = useState(false);
-
-  const handleCopy = async () => {
-    await navigator.clipboard.writeText(text);
-    setCopied(true);
-    trackCodeCopy("code-block");
-    setTimeout(() => setCopied(false), 2000);
-  };
-
-  return (
-    <button
-      onClick={handleCopy}
-      className="min-h-11 min-w-11 flex items-center justify-center rounded-lg hover:bg-white/10 transition-colors duration-150 bg-ink/80 backdrop-blur-sm"
-      aria-label="Copy code to clipboard"
-    >
-      {copied ? (
-        <Check className="w-4 h-4 text-accent-green" />
-      ) : (
-        <Copy className="w-4 h-4 text-cream-100/60 hover:text-cream-100 transition-colors" />
-      )}
-    </button>
-  );
-}
-
-function CodeBlock({
-  code,
-  language = "tsx",
-}: {
-  code: string;
-  language?: string;
-}) {
-  const lineCount = code.split("\n").length;
-
-  return (
-    <div className="relative group rounded-lg overflow-hidden">
-      <div className="absolute top-4 right-4 flex items-center gap-2 z-10">
-        <span className="text-xs text-cream-100/40 bg-ink/80 px-2 py-1 rounded backdrop-blur-sm">
-          {language}
-        </span>
-        <CopyButton text={code} />
-      </div>
-      <SyntaxHighlighter
-        language={language}
-        style={oneDark}
-        customStyle={{
-          margin: 0,
-          borderRadius: "0.5rem",
-          padding: "1.5rem",
-          paddingTop: "3.5rem",
-          fontSize: "0.875rem",
-          lineHeight: "1.6",
-          border: "1px solid rgba(255, 252, 248, 0.05)",
-        }}
-        showLineNumbers={lineCount > 10}
-        wrapLines={true}
-        wrapLongLines={false}
-      >
-        {code}
-      </SyntaxHighlighter>
-    </div>
-  );
-}
+import { CodeBlock, CopyButton } from "../components/code";
 
 export default function DevelopersPage() {
   return (
