+++ apps/developers/src/components/code/CopyButton.tsx
@@ -1,0 +1,34 @@
+"use client";
+
+import { useState } from "react";
+import { Check, Copy } from "lucide-react";
+import { trackCodeCopy } from "../../lib/analytics";
+
+interface CopyButtonProps {
+  text: string;
+}
+
+export function CopyButton({ text }: CopyButtonProps) {
+  const [copied, setCopied] = useState(false);
+
+  const handleCopy = async () => {
+    await navigator.clipboard.writeText(text);
+    setCopied(true);
+    trackCodeCopy("code-block");
+    setTimeout(() => setCopied(false), 2000);
+  };
+
+  return (
+    <button
+      onClick={handleCopy}
+      className="min-h-11 min-w-11 flex items-center justify-center rounded-lg hover:bg-white/10 transition-colors duration-150 bg-ink/80 backdrop-blur-sm"
+      aria-label="Copy code to clipboard"
+    >
+      {copied ? (
+        <Check className="w-4 h-4 text-accent-green" />
+      ) : (
+        <Copy className="w-4 h-4 text-cream-100/60 hover:text-cream-100 transition-colors" />
+      )}
+    </button>
+  );
+}
