+++ apps/developers/src/components/code/index.ts
@@ -1,0 +1,2 @@
+export { CodeBlock } from "./CodeBlock";
+export { CopyButton } from "./CopyButton";
